test_requirements.txt
pytest
pytest-pep8
pytest-pylint
pytest-cov
pytest-django
pytest-watch
mock
semantic_version
requests-mock==0.7.0

